pyproject.toml
[tool.poetry]
name = "hyperday"
version = "0.1.0"
description = ""
authors = ["Nickolsky <romb.intu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-novaclient = "^18.3.0"
python-dotenv = "^1.0.0"
flask = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
